Fixed Argument Loading Bug
Many of the bruteforcers loaded float-type arguments as integers, which
caused the values to be truncated. The problematic argument-loading
lines have been fixed.

As part of this fix, the type of the target speed found in the
LowSpeedRouteStruct 'TargetLog' has been changed from int to float. The
various references to this field have been updated accordingly, as has
the argument-loading line in LowSpeedRouter.cpp.

+++ GPUBullyImpactBruteforcer/GPUBullyImpactBruteforcer.cu
@@ -504,7 +504,7 @@ int main(int argc, char* argv[]) {
             i += 1;
         }
         else if (!strcmp(argv[i], "-vel")) {
-            bullyCentral.velBully = std::stoi(argv[i + 1]);
+            bullyCentral.velBully = std::stof(argv[i + 1]);
             i += 1;
         }
         else if (!strcmp(argv[i], "-o")) {
