+++ HighSpeedGPURouter/HiSpeedRouterGPUV2.cu
@@ -631,26 +631,26 @@ int main(int argc, char* argv[]) {
             exit(0);
         }
         else if (!strcmp(argv[i], "-ms")) {
-            firstSpeed = std::stoi(argv[i + 1]);
+            firstSpeed = std::stof(argv[i + 1]);
 
             i += 1;
         }
         else if (!strcmp(argv[i], "-ts")) {
-            targetSpeed = std::stoi(argv[i + 1]);
+            targetSpeed = std::stof(argv[i + 1]);
 
             i += 1;
         }
         else if (!strcmp(argv[i], "-cp")) {
-            cameraPosition[0] = std::stoi(argv[i + 1]);
-            cameraPosition[1] = std::stoi(argv[i + 2]);
-            cameraPosition[2] = std::stoi(argv[i + 3]);
+            cameraPosition[0] = std::stof(argv[i + 1]);
+            cameraPosition[1] = std::stof(argv[i + 2]);
+            cameraPosition[2] = std::stof(argv[i + 3]);
 
             i += 3;
         }
         else if (!strcmp(argv[i], "-fp")) {
-            firstPosition[0] = std::stoi(argv[i + 1]);
-            firstPosition[1] = std::stoi(argv[i + 2]);
-            firstPosition[2] = std::stoi(argv[i + 3]);
+            firstPosition[0] = std::stof(argv[i + 1]);
+            firstPosition[1] = std::stof(argv[i + 2]);
+            firstPosition[2] = std::stof(argv[i + 3]);
 
             i += 3;
         }
